+++ atcg_lib/src/Renderer/Buffer.cpp
@@ -4,14 +4,14 @@
 
 namespace atcg
 {
-    VertexBuffer::VertexBuffer(uint32_t size)
+    VertexBuffer::VertexBuffer(size_t size)
     {
         glGenBuffers(1, &_ID);
         glBindBuffer(GL_ARRAY_BUFFER, _ID);
         glBufferData(GL_ARRAY_BUFFER, size, nullptr, GL_DYNAMIC_DRAW);
     }
 
-    VertexBuffer::VertexBuffer(const void* data, uint32_t size)
+    VertexBuffer::VertexBuffer(const void* data, size_t size)
     {
         glGenBuffers(1, &_ID);
         glBindBuffer(GL_ARRAY_BUFFER, _ID);
